python-server/requirements.txt
Flask==1.1.1
setuptools >= 21.0.0
Pillow==6.2.1
